int and float parsing!
hell yes

+++ src/parsing/token.rs
@@ -1,7 +1,7 @@
 #[derive(Debug)]
 pub enum Token {
-    INT,
-    FLOAT,
+    INT(i32),
+    FLOAT(f32),
     ADD,
     SUBTRACT,
     MULTIPLY,
